fix crc

+++ ek-tm4c1294xl/project1/self/serial/ports_common.h
@@ -32,5 +32,5 @@ uint    MakeCrc16Bit31InBuff(uchar  bOffset, uint  wSize);
 
 uint    MakeCrc16Bit32InBuff(uchar  bOffset, uint  wSize);
 
-uchar   MakeCrc35OutBuff(uchar  bOffset, uint  wSize);
-uchar   MakeCrc35InBuff(uchar  bOffset, uint  wSize);
+uint    MakeCrc35OutBuff(uchar  bOffset, uint  wSize);
+uint    MakeCrc35InBuff(uchar  bOffset, uint  wSize);
